fix: не грузить винду и почистить в 427

+++ bin/update_tasks/0150002-dl_win-2-all.sh
@@ -2,6 +2,8 @@
 
 [ ! "`hostname | grep m1580-2`" ] && exit
 
+[ "`hostname | grep m1580-2-427`" ] && exit # не надо винду в 427
+
 cd /home/ && wget -nc https://sdo.1580.ru/some-files/Windows_10.ova && \
              wget -nc https://sdo.1580.ru/some-files/Windows_10.ova.md5 && \
              md5sum -c Windows_10.ova.md5 || (rm -f Windows_10.ova*; exit 1)
